+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- _Experimental feature_: add `FloggerAdapter` to use Flogger API to log events.
+
 ## [1.2.0] - 2020-10-24
 ### Added
 - Add a new module `autolog-aspectj` to allow usage of Autolog annotations in applications using AspectJ weaving.